@@ -4,7 +4,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
+
 
 import com.chistia007.cgpadom.databinding.ActivityLearnAppDevBinding;
 
@@ -16,25 +16,10 @@ public class LearnAppDev extends AppCompatActivity {
         binding=ActivityLearnAppDevBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        binding.btnGithub.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(LearnAppDev.this,GithubWebsiteActivity.class));
-            }
-        });
+        binding.btnGithub.setOnClickListener(view -> startActivity(new Intent(LearnAppDev.this,GithubWebsiteActivity.class)));
 
-        binding.btnLinkedin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(LearnAppDev.this,LinkdinWebsiteActivity.class));
-            }
-        });
+        binding.btnLinkedin.setOnClickListener(view -> startActivity(new Intent(LearnAppDev.this,LinkdinWebsiteActivity.class)));
 
-        binding.btnTwitter.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(LearnAppDev.this,TwitterWebsiteActivity.class));
-            }
-        });
+        binding.btnTwitter.setOnClickListener(view -> startActivity(new Intent(LearnAppDev.this,TwitterWebsiteActivity.class)));
     }
 }
