@@ -3,7 +3,6 @@ package com.chistia007.cgpadom;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
@@ -29,44 +28,25 @@ public class StartActivity extends AppCompatActivity {
 
         animation.setAnimationListener(new MyAnimationListener());
 
-        binding.btnCgpaCal.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(StartActivity.this,CalculateCgpaActivity.class).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP));
-            }
-        });
+        binding.btnCgpaCal.setOnClickListener(view -> startActivity(new Intent(StartActivity.this,CalculateCgpaActivity.class).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP)));
 
 
-        binding.btnBracuStudent.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                startActivity(new Intent(StartActivity.this,BracuStartActivity.class).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP));
-            }
-        });
+        binding.btnBracuStudent.setOnClickListener(view -> startActivity(new Intent(StartActivity.this,BracuStartActivity.class).addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP)));
 
-        binding.btnAboutUS.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                AlertDialog.Builder builder=new AlertDialog.Builder(StartActivity.this);
-                builder.setTitle(getString(R.string.app_name));
-                builder.setMessage("Designed and Developed with Love by Chisthia Khan From BRAC University\n"+
-                        "Check my page for more awesome applications");
-                builder.setNegativeButton("Dismiss", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
+        binding.btnAboutUS.setOnClickListener(view -> {
+            AlertDialog.Builder builder=new AlertDialog.Builder(StartActivity.this);
+            builder.setTitle(getString(R.string.app_name));
+            builder.setMessage("Designed and Developed with Love by Chisthia Khan From BRAC University\n"+
+                    "Check my page for more awesome applications");
+            builder.setNegativeButton("Dismiss", (dialog, which) -> {
 
-                    }
-                });
+            });
 
-                builder.setPositiveButton("Visit", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        Intent intent= new Intent(StartActivity.this, WebsiteActivity.class);
-                        startActivity(intent);
-                    }
-                });
-                builder.create().show();
-            }
+            builder.setPositiveButton("Visit", (dialog, which) -> {
+                Intent intent= new Intent(StartActivity.this, WebsiteActivity.class);
+                startActivity(intent);
+            });
+            builder.create().show();
         });
     }
 
