@@ -5,13 +5,12 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Toast;
 
 import com.chistia007.cgpadom.databinding.ActivityUpdateDashboardBinding;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
+
+import java.util.Objects;
 
 public class UpdateDashboard extends AppCompatActivity {
     ActivityUpdateDashboardBinding binding;
@@ -27,44 +26,35 @@ public class UpdateDashboard extends AppCompatActivity {
         mAuth = FirebaseAuth.getInstance();
 
 
-        binding.btnUpdateDash.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                binding.layVisibility1.setVisibility(View.VISIBLE);
-                binding.layVisibility2.setVisibility(View.VISIBLE);
+        binding.btnUpdateDash.setOnClickListener(view -> {
+            binding.layVisibility1.setVisibility(View.VISIBLE);
+            binding.layVisibility2.setVisibility(View.VISIBLE);
 
 
-                binding.btnUpdate.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        String credit=binding.edtCurrentCredit.getText().toString();
-                        String cgpa=binding.edtCurrentCgpa.getText().toString();
-                        db.collection("Users").document(mAuth.getCurrentUser().getUid())
-                                .update(
-                                        "cgpa", cgpa,
-                                        "credit", credit
-                                );
-                        Intent intent= new Intent(UpdateDashboard.this, CgpaBracuCalculatorActivity.class);
-                        startActivity(intent);
-                    }
-                });
-            }
+            binding.btnUpdate.setOnClickListener(view1 -> {
+                String credit= Objects.requireNonNull(binding.edtCurrentCredit.getText()).toString();
+                String cgpa= Objects.requireNonNull(binding.edtCurrentCgpa.getText()).toString();
+                db.collection("Users").document(Objects.requireNonNull(mAuth.getCurrentUser()).getUid())
+                        .update(
+                                "cgpa", cgpa,
+                                "credit", credit
+                        );
+                Intent intent= new Intent(UpdateDashboard.this, CgpaBracuCalculatorActivity.class);
+                startActivity(intent);
+            });
         });
 
 
 
 
-        binding.btnResetDash.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                db.collection("Users").document(mAuth.getCurrentUser().getUid())
-                        .update(
-                                "cgpa", "0",
-                                "credit", "0"
-                        );
-                Intent intent= new Intent(UpdateDashboard.this, CgpaBracuCalculatorActivity.class);
-                startActivity(intent);
-            }
+        binding.btnResetDash.setOnClickListener(view -> {
+            db.collection("Users").document(Objects.requireNonNull(mAuth.getCurrentUser()).getUid())
+                    .update(
+                            "cgpa", "0",
+                            "credit", "0"
+                    );
+            Intent intent= new Intent(UpdateDashboard.this, CgpaBracuCalculatorActivity.class);
+            startActivity(intent);
         });
     }
 }
