@@ -1,26 +1,16 @@
 package com.chistia007.cgpadom;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.app.ProgressDialog;
 import android.content.Intent;
-import android.nfc.Tag;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
 
 import com.chistia007.cgpadom.databinding.ActivityRegisterBinding;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 
@@ -71,51 +61,37 @@ public class RegisterActivity extends AppCompatActivity {
                         Toast.makeText(RegisterActivity.this, "Password too short", Toast.LENGTH_SHORT).show();
                     }
                     else{
-                        registerUser(txtName,txtEmail,txtPassword);
+                        registerUser(txtEmail,txtPassword);
                     }
                 }
 
             }
 
-            private void registerUser(String name, String email, String password) {
+            private void registerUser(String email, String password) {
                 progressDialog.setTitle("Please wait...");
                 progressDialog.show();
-                mAuth.createUserWithEmailAndPassword(email,password).addOnSuccessListener(new OnSuccessListener<AuthResult>() {
-                    @Override
-                    public void onSuccess(AuthResult authResult) {
-                        Toast.makeText(RegisterActivity.this, "Registered Successfully", Toast.LENGTH_SHORT).show();
-                        FirebaseFirestore db = FirebaseFirestore.getInstance();
-                        Map<String,String> user=new HashMap<>();
-                        user.put("name",binding.edtName.getText().toString());
-                        user.put("email",binding.edtEmail.getText().toString());
-                        user.put("password",binding.edtPassword.getText().toString());
-                        user.put("credit","0");
-                        user.put("cgpa","0");
+                mAuth.createUserWithEmailAndPassword(email,password).addOnSuccessListener(authResult -> {
+                    Toast.makeText(RegisterActivity.this, "Registered Successfully", Toast.LENGTH_SHORT).show();
+                    FirebaseFirestore db = FirebaseFirestore.getInstance();
+                    Map<String,String> user=new HashMap<>();
+                    user.put("name",binding.edtName.getText().toString());
+                    user.put("email",binding.edtEmail.getText().toString());
+                    user.put("password",binding.edtPassword.getText().toString());
+                    user.put("credit","0");
+                    user.put("cgpa","0");
 
-                        CollectionReference col = db.collection("Users");
+                    CollectionReference col = db.collection("Users");
 
-                        col.document(Objects.requireNonNull(mAuth.getCurrentUser()).getUid()).set(user).addOnSuccessListener(new OnSuccessListener<Void>() {
-                            @Override
-                            public void onSuccess(Void aVoid) {
+                    col.document(Objects.requireNonNull(mAuth.getCurrentUser()).getUid()).set(user).addOnSuccessListener(aVoid -> {
 
-                            }
-                        })
-                                .addOnFailureListener(new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        Toast.makeText(RegisterActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                                    }
-                                });
-                        progressDialog.dismiss();
-                        startActivity(new Intent(RegisterActivity.this,CgpaBracuCalculatorActivity.class));
+                    })
+                            .addOnFailureListener(e -> Toast.makeText(RegisterActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show());
+                    progressDialog.dismiss();
+                    startActivity(new Intent(RegisterActivity.this,CgpaBracuCalculatorActivity.class));
 
-                    }
-                }).addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        progressDialog.dismiss();
-                        Toast.makeText(RegisterActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                    }
+                }).addOnFailureListener(e -> {
+                    progressDialog.dismiss();
+                    Toast.makeText(RegisterActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show();
                 });
             }
         });
